+++ changelog.md
@@ -14,4 +14,5 @@
 
 ## Next
 
+* Remove redundant native-or-bluebird dependency
 * Update dev dependencies